CE->EE merge before 8.13.0-rc1

Conflicts resolved in the first merge:

```
UU app/models/group.rb
UU app/models/member.rb
UU app/models/project_team.rb
UU app/services/merge_requests/update_service.rb
UU app/services/notification_service.rb
UU app/views/shared/_sort_dropdown.html.haml
UU doc/api/oauth2.md
UU lib/api/members.rb
UU spec/models/merge_request_spec.rb
AA spec/views/ci/lints/show.html.haml_spec.rb
UU app/views/projects/commits/_head.html.haml
```

Conflicts resolved in the second merge:

```
UU app/assets/javascripts/groups_select.js
UU app/assets/javascripts/labels_select.js
UU app/controllers/concerns/membership_actions.rb
UU app/controllers/groups/group_members_controller.rb
UU app/helpers/selects_helper.rb
UU app/views/admin/dashboard/index.html.haml
UU app/views/shared/issuable/_form.html.haml
UU app/views/shared/issuable/_label_dropdown.html.haml
UU app/views/shared/issuable/_sidebar.html.haml
UU config/routes.rb
UU spec/services/merge_requests/merge_service_spec.rb
UU spec/services/system_note_service_spec.rb
```

See merge request !772

+++ app/models/merge_request.rb
@@ -160,6 +160,20 @@ class MergeRequest < ActiveRecord::Base
     where("merge_requests.id IN (#{union.to_sql})")
   end
 
+  WIP_REGEX = /\A\s*(\[WIP\]\s*|WIP:\s*|WIP\s+)+\s*/i.freeze
+
+  def self.work_in_progress?(title)
+    !!(title =~ WIP_REGEX)
+  end
+
+  def self.wipless_title(title)
+    title.sub(WIP_REGEX, "")
+  end
+
+  def self.wip_title(title)
+    work_in_progress?(title) ? title : "WIP: #{title}"
+  end
+
   def to_reference(from_project = nil)
     reference = "#{self.class.reference_prefix}#{iid}"
 
@@ -410,14 +424,16 @@ class MergeRequest < ActiveRecord::Base
     @closed_event ||= target_project.events.where(target_id: self.id, target_type: "MergeRequest", action: Event::CLOSED).last
   end
 
-  WIP_REGEX = /\A\s*(\[WIP\]\s*|WIP:\s*|WIP\s+)+\s*/i.freeze
-
   def work_in_progress?
-    !!(title =~ WIP_REGEX)
+    self.class.work_in_progress?(title)
   end
 
   def wipless_title
-    self.title.sub(WIP_REGEX, "")
+    self.class.wipless_title(self.title)
+  end
+
+  def wip_title
+    self.class.wip_title(self.title)
   end
 
   def mergeable?(skip_ci_check: false)
@@ -528,9 +544,13 @@ class MergeRequest < ActiveRecord::Base
   # `MergeRequestsClosingIssues` model. This is a performance optimization.
   # Calculating this information for a number of merge requests requires
   # running `ReferenceExtractor` on each of them separately.
+  # This optimization does not apply to issues from external sources.
   def cache_merge_request_closes_issues!(current_user = self.author)
+    return if project.has_external_issue_tracker?
+
     transaction do
       self.merge_requests_closing_issues.delete_all
+
       closes_issues(current_user).each do |issue|
         self.merge_requests_closing_issues.create!(issue: issue)
       end
